refactor Moved functions from base file further
