+++ app/src/util/tools/index.ts
@@ -15,3 +15,5 @@ export * from './string_manipulation'
 export * from './dates'
 export * from './utils'
 export * from './web3'
+export * from './formatting'
+export * from './other'
